third_party/blink/web_tests/external/wpt/html/canvas/offscreen/the-canvas-state/2d.state.saverestore.lineWidth.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.state.saverestore.lineWidth</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.state.saverestore.lineWidth</h1>

<script>

test(t => {
  const canvas = new OffscreenCanvas(300, 150);
  const ctx = canvas.getContext('2d');

  // Test that restore() undoes any modifications.
  var old = ctx.lineWidth;
  ctx.save();
  ctx.lineWidth = 0.5;
  ctx.restore();
  _assertSame(ctx.lineWidth, old, "ctx.lineWidth", "old");

  // Also test that save() doesn't modify the values.
  ctx.lineWidth = 0.5;
  old = ctx.lineWidth;
  // We're not interested in failures caused by get(set(x)) != x (e.g.
  // from rounding), so compare against `old` instead of against 0.5.
  ctx.save();
  _assertSame(ctx.lineWidth, old, "ctx.lineWidth", "old");
  ctx.restore();
}, "save()/restore() works for lineWidth, with a canvas size of (300, 150)");

test(t => {
  const canvas = new OffscreenCanvas(0, 0);
  const ctx = canvas.getContext('2d');

  // Test that restore() undoes any modifications.
  var old = ctx.lineWidth;
  ctx.save();
  ctx.lineWidth = 0.5;
  ctx.restore();
  _assertSame(ctx.lineWidth, old, "ctx.lineWidth", "old");

  // Also test that save() doesn't modify the values.
  ctx.lineWidth = 0.5;
  old = ctx.lineWidth;
  // We're not interested in failures caused by get(set(x)) != x (e.g.
  // from rounding), so compare against `old` instead of against 0.5.
  ctx.save();
  _assertSame(ctx.lineWidth, old, "ctx.lineWidth", "old");
  ctx.restore();
}, "save()/restore() works for lineWidth, with a canvas size of (0, 0)");

</script>
